requirements.txt
pyyaml
unidecode